+++ src/app.tsx
@@ -3,26 +3,14 @@ import './styles/global.css';
 import Tab from './components/Tab';
 import Tabs from './components/Tabs';
 import Badge from './components/Badge';
-import React, { useEffect, useState } from 'preact/compat';
-import {
-  ElementType,
-  GameData,
-  Tower,
-  Wizard,
-  WizardCounts,
-  WizardCollection,
-  WizardDispatchData,
-} from './types';
+import { useEffect, useState } from 'preact/compat';
+import { ElementType, GameData, Tower, WizardCollection } from './types';
 import { eventEmitter } from './events/EventEmitter';
-import fire_wizard_src from '../src/assets/wizards/wizard_red.png';
-import water_wizard_src from '../src/assets/wizards/wizard_blue.png';
-import earth_wizard_src from '../src/assets/wizards/wizard_brown.png';
-import air_wizard_src from '../src/assets/wizards/wizard_white.png';
 import { Tile } from './systems/world_generation/Tile';
-import { eventImages } from './app-utils/image-maps';
-import { WizardProfile } from './components/WizardProfile';
 import { WizardsTab } from './components/WizardsTab';
 import { EmpiresTab } from './components/EmpiresTab';
+import { TowerTab } from './components/TowerTab';
+import { TilesTab } from './components/TilesTab';
 
 export function App() {
   const [gameState, setGameState] = useState<GameData>(null);
@@ -174,211 +162,3 @@ export function App() {
     </div>
   );
 }
-
-const TowerTab: React.FC<{
-  wizards: WizardCollection;
-  tower: Tower;
-  playerGold: number;
-  buy: (element: ElementType) => void;
-}> = (props) => {
-  const { wizards, tower, playerGold, buy } = props;
-
-  return (
-    <div className="flex flex-col p-2 max-h-[90vh] overflow-auto">
-      <h2 className="text-2xl text-red-500 pb-2">Tower</h2>
-      {['fire', 'water', 'earth', 'air'].map((element) => (
-        <div className="py-2">
-          <WizardShopPanel
-            elementType={element}
-            playerGold={playerGold}
-            wizardTypeCount={wizards[element as keyof WizardCollection].length}
-            wizardTypeCapacity={
-              tower.wizardCapacities[element as keyof WizardCounts]
-            }
-            wizardCost={
-              tower.baseWizardCost +
-              tower.perExtraWizardCost *
-                (wizards[element as keyof WizardCollection].length - 1)
-            }
-            buy={buy}
-          />
-        </div>
-      ))}
-    </div>
-  );
-};
-
-const TilesTab: React.FC<{
-  currentTile: Tile | null;
-  wizards: WizardCollection;
-}> = (props) => {
-  const { currentTile, wizards } = props;
-
-  const handleDispatchWizard = (targetWizard: Wizard, targetTile: Tile) => {
-    let wizardDispatchData: WizardDispatchData = {
-      wizard: targetWizard,
-      tile: targetTile,
-    };
-
-    eventEmitter.emit('dispatch-wizard', wizardDispatchData);
-  };
-
-  if (currentTile === null)
-    return <p className="text-white">No tile selected</p>;
-
-  return (
-    <div className="flex flex-col max-h-[90vh] h-full overflow-auto">
-      <h2 className="text-2xl text-purple-400 pb-2">
-        {`(${currentTile.coordinates.x}, ${currentTile.coordinates.y}) ${currentTile.terrainData.name}`}
-      </h2>
-      <p className="text-white">
-        {currentTile.terrainData.is_walkable ? 'Walkable' : 'Not walkable'}
-      </p>
-      <p className="text-white pt-2">
-        Structure:{' '}
-        {currentTile.structureData ? currentTile.structureData.name : 'None'}
-      </p>
-      <div className="text-white py-2">
-        Event:{' '}
-        {currentTile.currentEvent
-          ? currentTile.currentEvent.worldEventSettings.type
-          : 'None'}
-        {currentTile.currentEvent ? (
-          <img
-            src={eventImages[currentTile.currentEvent.worldEventSettings.type]}
-            width={'100%'}
-            className={'py-2'}
-          />
-        ) : null}
-      </div>
-      <p className="text-white">Wizards here: </p>
-      <p className="text-white">(none)</p>
-      <p className="text-white pt-2">Send a wizard here: </p>
-      {wizards.air.map((wizard) => (
-        <WizardProfile
-          key={wizard.name}
-          wizard={wizard}
-          clickedCallback={() => {
-            handleDispatchWizard(wizard, currentTile);
-          }}
-        />
-      ))}
-      {wizards.earth.map((wizard) => (
-        <WizardProfile
-          key={wizard.name}
-          wizard={wizard}
-          clickedCallback={() => {
-            handleDispatchWizard(wizard, currentTile);
-          }}
-        />
-      ))}
-      {wizards.fire.map((wizard) => (
-        <WizardProfile
-          key={wizard.name}
-          wizard={wizard}
-          clickedCallback={() => {
-            handleDispatchWizard(wizard, currentTile);
-          }}
-        />
-      ))}
-      {wizards.water.map((wizard) => (
-        <WizardProfile
-          key={wizard.name}
-          wizard={wizard}
-          clickedCallback={() => {
-            handleDispatchWizard(wizard, currentTile);
-          }}
-        />
-      ))}
-    </div>
-  );
-};
-
-const WizardShopPanel: React.FC<{
-  elementType: string;
-  playerGold: number;
-  wizardTypeCount: number;
-  wizardTypeCapacity: number;
-  wizardCost: number;
-  buy: (element: ElementType) => void;
-}> = (props) => {
-  const {
-    elementType,
-    playerGold,
-    wizardTypeCount,
-    wizardTypeCapacity,
-    wizardCost,
-    buy,
-  } = props;
-
-  let bgColor;
-  let btnColor;
-  let btnHoverColor;
-  let wizardImgSrc;
-
-  switch (elementType) {
-    case 'fire':
-      bgColor = 'bg-red-400';
-      btnColor = 'bg-red-600';
-      btnHoverColor = 'hover:bg-red-900';
-      wizardImgSrc = fire_wizard_src;
-      break;
-    case 'water':
-      bgColor = 'bg-blue-400';
-      btnColor = 'bg-blue-600';
-      btnHoverColor = 'hover:bg-blue-900';
-      wizardImgSrc = water_wizard_src;
-      break;
-    case 'earth':
-      bgColor = 'bg-amber-400';
-      btnColor = 'bg-amber-600';
-      btnHoverColor = 'hover:bg-amber-900';
-      wizardImgSrc = earth_wizard_src;
-      break;
-    case 'air':
-      bgColor = 'bg-slate-400';
-      btnColor = 'bg-gray-600';
-      btnHoverColor = 'hover:bg-gray-900';
-      wizardImgSrc = air_wizard_src;
-      break;
-    default:
-      bgColor = 'bg-gray-400';
-      btnColor = 'bg-gray-700';
-      btnHoverColor = 'hover:bg-gray-900';
-      wizardImgSrc = fire_wizard_src;
-  }
-
-  return (
-    <div
-      className={`${bgColor} w-full h-48 rounded-lg p-4 flex flex-row justify-between`}
-    >
-      <div className="w-1/3 h-full">
-        <img
-          src={`${wizardImgSrc}`}
-          alt={`${elementType} wizard`}
-          className="w-full h-full object-contain"
-        />
-      </div>
-      <div className="w-1/2 h-full flex flex-col justify-center items-center">
-        <div className="flex flex-col justify-center items-center">
-          <p className="text-white">Capacity:</p>
-          <p className="text-white">
-            {wizardTypeCount} of {wizardTypeCapacity}
-          </p>
-        </div>
-        <button
-          className={`${btnColor} text-white px-3 py-4 rounded-md shadow-md ${btnHoverColor} hover:text-gray disabled:bg-[#333333ff] `}
-          disabled={
-            wizardTypeCount >= wizardTypeCapacity || playerGold < wizardCost
-          }
-          onClick={() => {
-            buy(elementType as ElementType);
-          }}
-        >
-          <p>Buy {elementType} Wizard:</p>
-          <p>{wizardCost} Gold</p>
-        </button>
-      </div>
-    </div>
-  );
-};
